Update form done + file structure + modal
