+++ frontend/src/Components/App.jsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import PlayerList from './Player/PlayerList';
 import PlayerSingle from './Player/PlayerSingle';
-import PlayerForm from './Player/PlayerForm';
+import AddPlayerForm from './Interaction/AddPlayerForm';
+import Modal from './Modal/Modal';
 
 const App = () => { 
 
   const [players, setPlayers] = useState([]);
   const [currentPlayer, setCurrentPlayer] = useState({});
+  const [shouldShowModal, setShouldShowModal] = useState(false)
 
   const updateCurrentPlayer = (item) => {
     setCurrentPlayer(item);
@@ -30,11 +32,13 @@ const App = () => {
   }, [players]);
     
     return (
+      <>
       <div className="container-fluid">
         <div className="row">
           <nav>
-            <div className="nav-wrapper blue darken-1">
-              <a href="/" className="brand-logo">Soccer Management</a>
+            <div className=" blue darken-1" style={{display: "flex", justifyContent: "space-between", padding: "5px 15px 5px 15px"}}>
+              <a href="/" style={{fontSize: "35px"}}>Soccer Management</a>
+              <div onClick={() => setShouldShowModal(!shouldShowModal)} style={{fontSize: "20px", textAlign: "center", cursor: "pointer"}}><span style={{fontSize: "30px", }}>+</span> Add player</div>
             </div>
           </nav>
         </div>
@@ -42,12 +46,13 @@ const App = () => {
           <div className="col s3"><PlayerList players={players}
             updateCurrentPlayer={updateCurrentPlayer}/>
           </div>
-          <div className="col s9"><PlayerSingle player={currentPlayer}/></div>
-        </div>
-        <div className="row">
-          <div className="col s12"><PlayerForm /></div>
+          <div className="col s9"><PlayerSingle player={currentPlayer} setCurrentPlayer={setCurrentPlayer} /></div>
         </div>
       </div>
+      <Modal shouldShow={shouldShowModal} onRequestClose={() => setShouldShowModal(false)} >
+        <AddPlayerForm />
+      </Modal>  
+    </>
     );
 }
