--- frontend/src/Components/Player/PlayerSingle.js
@@ -1,27 +1,0 @@
-import React from 'react';
-
-const PlayerSingle = (props) => {
-
-    return ( 
-    <div className="row">
-        <div className="col s12">
-            <div className="card">
-                <div className="card-image">
-                    <img src="soccer.jpeg" alt='bild' />
-                    <span className="card-title">{props.player.firstName} {props.player.lastName}</span>
-                    <a href="/" className="btn-floating halfway-fab ">Update</a>
-                </div>
-                <div className="card-content">
-                    <p>Phone: {props.player.phone} - Email: {props.player.email}</p>
-                    <p>Strength: {props.player.strength} - Endurance: {props.player.endurance}</p>
-                </div>
-                <div className='card-action'>
-                    Team: {props.player.team}
-                </div>
-            </div>
-        </div>
-    </div>
-    );
-};
-
-export default PlayerSingle;
